golang practice with proj euler
not sure whether to use if-conditionals or error-handling when
potentially accessing an index out of range of the slice/array.  or use
a map?
(also man i miss python's conciseness)
I think it works now (? possibly ?) but it is mostly likely not proper
go, probably because I lost track of which variables are pointers and
which are objects.